+++ shared/src/commonTest/kotlin/org/hyperskill/step_quiz_code_blanks/presentation/StepQuizCodeBlanksCreateReplyTest.kt
@@ -1,65 +1,31 @@
-package org.hyperskill.step_quiz_code_blanks
+package org.hyperskill.step_quiz_code_blanks.presentation
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertFalse
-import kotlin.test.assertNull
-import kotlin.test.assertTrue
 import org.hyperskill.app.step.domain.model.Block
 import org.hyperskill.app.step.domain.model.Step
 import org.hyperskill.app.step_quiz_code_blanks.domain.model.CodeBlock
 import org.hyperskill.app.step_quiz_code_blanks.domain.model.CodeBlockChild
 import org.hyperskill.app.step_quiz_code_blanks.domain.model.Suggestion
 import org.hyperskill.app.step_quiz_code_blanks.presentation.StepQuizCodeBlanksFeature
-import org.hyperskill.app.step_quiz_code_blanks.presentation.activeCodeBlockIndex
 import org.hyperskill.app.step_quiz_code_blanks.presentation.createReply
-import org.hyperskill.app.step_quiz_code_blanks.presentation.isVariableSuggestionsAvailable
 import org.hyperskill.app.submissions.domain.model.Reply
 import org.hyperskill.step.domain.model.stub
 
-class StepQuizCodeBlanksStateExtensionsTest {
+class StepQuizCodeBlanksCreateReplyTest {
     companion object {
         private const val REPLY_CODE_LANGUAGE = "python3"
         private const val REPLY_CODE_PREFIX = "# solved with code blanks\n"
     }
 
-    @Test
-    fun `activeCodeBlockIndex should return null if no active code block`() {
-        val state = stubContentState(
-            codeBlocks = listOf(
-                CodeBlock.Blank(isActive = false, suggestions = emptyList()),
-                CodeBlock.Print(
-                    children = listOf(
-                        CodeBlockChild.SelectSuggestion(
-                            isActive = false,
-                            suggestions = emptyList(),
-                            selectedSuggestion = null
-                        )
-                    )
-                )
+    private val step = Step.stub(
+        id = 1,
+        block = Block.stub(
+            options = Block.Options(
+                codeTemplates = mapOf(REPLY_CODE_LANGUAGE to "# put your python code here")
             )
         )
-        assertNull(state.activeCodeBlockIndex())
-    }
-
-    @Test
-    fun `activeCodeBlockIndex should return index of the active code block`() {
-        val state = stubContentState(
-            codeBlocks = listOf(
-                CodeBlock.Blank(isActive = false, suggestions = emptyList()),
-                CodeBlock.Print(
-                    children = listOf(
-                        CodeBlockChild.SelectSuggestion(
-                            isActive = true,
-                            suggestions = emptyList(),
-                            selectedSuggestion = null
-                        )
-                    )
-                )
-            )
-        )
-        assertEquals(1, state.activeCodeBlockIndex())
-    }
+    )
 
     @Test
     fun `createReply should return Reply with code from code blocks and language from step options`() {
@@ -75,7 +41,7 @@ class StepQuizCodeBlanksStateExtensionsTest {
             ),
             CodeBlock.Blank(isActive = true, suggestions = emptyList())
         )
-        val state = stubContentState(codeBlocks = codeBlocks)
+        val state = StepQuizCodeBlanksFeature.State.Content.stub(step = step, codeBlocks = codeBlocks)
 
         val expectedReply = Reply.code(
             code = buildString {
@@ -115,7 +81,7 @@ class StepQuizCodeBlanksStateExtensionsTest {
                 )
             ),
         )
-        val state = stubContentState(codeBlocks = codeBlocks)
+        val state = StepQuizCodeBlanksFeature.State.Content.stub(step = step, codeBlocks = codeBlocks)
 
         val expectedReply = Reply.code(
             code = buildString {
@@ -233,7 +199,7 @@ class StepQuizCodeBlanksStateExtensionsTest {
                 )
             )
         )
-        val state = stubContentState(codeBlocks = codeBlocks)
+        val state = StepQuizCodeBlanksFeature.State.Content.stub(step = step, codeBlocks = codeBlocks)
 
         val expectedReply = Reply.code(
             code = buildString {
@@ -368,7 +334,7 @@ class StepQuizCodeBlanksStateExtensionsTest {
                 )
             )
         )
-        val state = stubContentState(codeBlocks = codeBlocks)
+        val state = StepQuizCodeBlanksFeature.State.Content.stub(step = step, codeBlocks = codeBlocks)
 
         val expectedReply = Reply.code(
             code = buildString {
@@ -446,7 +412,7 @@ class StepQuizCodeBlanksStateExtensionsTest {
                 )
             )
         )
-        val state = stubContentState(codeBlocks = codeBlocks)
+        val state = StepQuizCodeBlanksFeature.State.Content.stub(step = step, codeBlocks = codeBlocks)
 
         val expectedReply = Reply.code(
             code = buildString {
@@ -597,7 +563,7 @@ class StepQuizCodeBlanksStateExtensionsTest {
                 )
             )
         )
-        val state = stubContentState(codeBlocks = codeBlocks)
+        val state = StepQuizCodeBlanksFeature.State.Content.stub(step = step, codeBlocks = codeBlocks)
 
         val expectedReply = Reply.code(
             code = buildString {
@@ -617,49 +583,4 @@ class StepQuizCodeBlanksStateExtensionsTest {
 
         assertEquals(expectedReply, state.createReply())
     }
-
-    @Test
-    fun `isVariableSuggestionsAvailable should return true if variable suggestions are available`() {
-        val step = Step.stub(
-            id = 1,
-            block = Block.stub(options = Block.Options(codeBlanksVariables = listOf("a", "b")))
-        )
-        val state = stubContentState(
-            step = step,
-            codeBlocks = emptyList()
-        )
-
-        assertTrue(state.isVariableSuggestionsAvailable)
-    }
-
-    @Test
-    fun `isVariableSuggestionsAvailable should return false if variable suggestions are not available`() {
-        listOf(null, emptyList<String>()).forEach { codeBlanksVariables ->
-            val step = Step.stub(
-                id = 1,
-                block = Block.stub(options = Block.Options(codeBlanksVariables = codeBlanksVariables))
-            )
-            val state = stubContentState(
-                step = step,
-                codeBlocks = emptyList()
-            )
-
-            assertFalse(state.isVariableSuggestionsAvailable)
-        }
-    }
-
-    private fun stubContentState(
-        step: Step = Step.stub(id = 1).copy(
-            block = Block.stub(
-                options = Block.Options(
-                    codeTemplates = mapOf(REPLY_CODE_LANGUAGE to "# put your python code here")
-                )
-            )
-        ),
-        codeBlocks: List<CodeBlock>
-    ): StepQuizCodeBlanksFeature.State.Content =
-        StepQuizCodeBlanksFeature.State.Content(
-            step = step,
-            codeBlocks = codeBlocks
-        )
 }
